Fix configuration for running tests
For a development it's now possible to run `npm start` which
creates test bundle and starts Test'em to run tests. The global
Scent variable is defined using webpack ProvidePlugin and points
to _src/scent.coffee_ so upon changes a bundle is updated and
browser with tests is reloaded.

For a CI or just single run of tests simply run `npm test`. It will
compile files into _lib_ folder using Coffee and Babel and these
files are then used in tests by setting global Scent variable
pointing to _lib/scent.js_. This behavior is defined in
_test/run.coffee_.

There is also temporary repo for sinon-chai while waiting for sinon
to be released. The sinon-chai peerDependency doesn't satisfy version
2.0.0-pre of sinon.